+++ packages/connectivity/connectivity_platform_interface/lib/connectivity_platform_interface.dart
@@ -9,39 +9,6 @@ import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'method_channel_connectivity.dart';
 
-enum ConnectivityResult {
-  /// WiFi: Device connected via Wi-Fi
-  wifi,
-
-  /// Mobile: Device connected to cellular network
-  mobile,
-
-  /// None: Device not connected to any network
-  none
-}
-
-
-/// The status of the location service authorization.
-enum LocationAuthorizationStatus {
-  /// The authorization of the location service is not determined.
-  notDetermined,
-
-  /// This app is not authorized to use location.
-  restricted,
-
-  /// User explicitly denied the location service.
-  denied,
-
-  /// User authorized the app to access the location at any time.
-  authorizedAlways,
-
-  /// User authorized the app to access the location when the app is visible to them.
-  authorizedWhenInUse,
-
-  /// Status unknown.
-  unknown
-}
-
 /// The interface that implementations of connectivity must implement.
 ///
 /// Platform implementations should extend this class rather than implement it as `connectivity`
@@ -62,37 +29,6 @@ abstract class ConnectivityPlatform extends PlatformInterface {
   /// Defaults to [MethodChannelConnectivity].
   static ConnectivityPlatform get instance => _instance;
 
-  @protected
-  ConnectivityResult parseConnectivityResult(String state) {
-    switch (state) {
-      case 'wifi':
-        return ConnectivityResult.wifi;
-      case 'mobile':
-        return ConnectivityResult.mobile;
-      case 'none':
-      default:
-        return ConnectivityResult.none;
-    }
-  }
-  @protected
-  LocationAuthorizationStatus convertLocationStatusString(String result) {
-    switch (result) {
-      case 'notDetermined':
-        return LocationAuthorizationStatus.notDetermined;
-      case 'restricted':
-        return LocationAuthorizationStatus.restricted;
-      case 'denied':
-        return LocationAuthorizationStatus.denied;
-      case 'authorizedAlways':
-        return LocationAuthorizationStatus.authorizedAlways;
-      case 'authorizedWhenInUse':
-        return LocationAuthorizationStatus.authorizedWhenInUse;
-      default:
-        return LocationAuthorizationStatus.unknown;
-    }
-  }
-
-
   /// Platform-specific plugins should set this with their own platform-specific
   /// class that extends [ConnectivityPlatform] when they register themselves.
   // TODO(amirh): Extract common platform interface logic.
@@ -102,11 +38,11 @@ abstract class ConnectivityPlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  Future<ConnectivityResult> checkConnectivity();
+  Future<String> checkConnectivity();
   Future<String> getWifiName();
   Future<String> getWifiBSSID();
   Future<String> getWifiIP();
-  Future<LocationAuthorizationStatus> requestLocationServiceAuthorization(
+  Future<String> requestLocationServiceAuthorization(
       {bool requestAlwaysLocationUsage = false});
-  Future<LocationAuthorizationStatus> getLocationServiceAuthorization();
+  Future<String> getLocationServiceAuthorization();
 }
