+++ packages/connectivity/connectivity_platform_interface/test/method_channel_connectivity_test.dart
@@ -1,289 +1,289 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
+// // Copyright 2017 The Chromium Authors. All rights reserved.
+// // Use of this source code is governed by a BSD-style license that can be
+// // found in the LICENSE file.
 
-import 'package:mockito/mockito.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+// import 'package:mockito/mockito.dart';
+// import 'package:flutter/services.dart';
+// import 'package:flutter_test/flutter_test.dart';
+// import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-import 'package:url_launcher_platform_interface/method_channel_url_launcher.dart';
-import 'package:url_launcher_platform_interface/url_launcher_platform_interface.dart';
+// import 'package:url_launcher_platform_interface/method_channel_url_launcher.dart';
+// import 'package:url_launcher_platform_interface/url_launcher_platform_interface.dart';
 
-void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
+// void main() {
+//   TestWidgetsFlutterBinding.ensureInitialized();
 
-  group('$UrlLauncherPlatform', () {
-    test('$MethodChannelUrlLauncher() is the default instance', () {
-      expect(UrlLauncherPlatform.instance,
-          isInstanceOf<MethodChannelUrlLauncher>());
-    });
+//   group('$UrlLauncherPlatform', () {
+//     test('$MethodChannelUrlLauncher() is the default instance', () {
+//       expect(UrlLauncherPlatform.instance,
+//           isInstanceOf<MethodChannelUrlLauncher>());
+//     });
 
-    test('Cannot be implemented with `implements`', () {
-      expect(() {
-        UrlLauncherPlatform.instance = ImplementsUrlLauncherPlatform();
-      }, throwsA(isInstanceOf<AssertionError>()));
-    });
+//     test('Cannot be implemented with `implements`', () {
+//       expect(() {
+//         UrlLauncherPlatform.instance = ImplementsUrlLauncherPlatform();
+//       }, throwsA(isInstanceOf<AssertionError>()));
+//     });
 
-    test('Can be mocked with `implements`', () {
-      final UrlLauncherPlatformMock mock = UrlLauncherPlatformMock();
-      UrlLauncherPlatform.instance = mock;
-    });
+//     test('Can be mocked with `implements`', () {
+//       final UrlLauncherPlatformMock mock = UrlLauncherPlatformMock();
+//       UrlLauncherPlatform.instance = mock;
+//     });
 
-    test('Can be extended', () {
-      UrlLauncherPlatform.instance = ExtendsUrlLauncherPlatform();
-    });
-  });
+//     test('Can be extended', () {
+//       UrlLauncherPlatform.instance = ExtendsUrlLauncherPlatform();
+//     });
+//   });
 
-  group('$MethodChannelUrlLauncher', () {
-    const MethodChannel channel =
-        MethodChannel('plugins.flutter.io/url_launcher');
-    final List<MethodCall> log = <MethodCall>[];
-    channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      log.add(methodCall);
-    });
+//   group('$MethodChannelUrlLauncher', () {
+//     const MethodChannel channel =
+//         MethodChannel('plugins.flutter.io/url_launcher');
+//     final List<MethodCall> log = <MethodCall>[];
+//     channel.setMockMethodCallHandler((MethodCall methodCall) async {
+//       log.add(methodCall);
+//     });
 
-    final MethodChannelUrlLauncher launcher = MethodChannelUrlLauncher();
+//     final MethodChannelUrlLauncher launcher = MethodChannelUrlLauncher();
 
-    tearDown(() {
-      log.clear();
-    });
+//     tearDown(() {
+//       log.clear();
+//     });
 
-    test('canLaunch', () async {
-      await launcher.canLaunch('http://example.com/');
-      expect(
-        log,
-        <Matcher>[
-          isMethodCall('canLaunch', arguments: <String, Object>{
-            'url': 'http://example.com/',
-          })
-        ],
-      );
-    });
+//     test('canLaunch', () async {
+//       await launcher.canLaunch('http://example.com/');
+//       expect(
+//         log,
+//         <Matcher>[
+//           isMethodCall('canLaunch', arguments: <String, Object>{
+//             'url': 'http://example.com/',
+//           })
+//         ],
+//       );
+//     });
 
-    test('launch', () async {
-      await launcher.launch(
-        'http://example.com/',
-        useSafariVC: true,
-        useWebView: false,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: false,
-        headers: const <String, String>{},
-      );
-      expect(
-        log,
-        <Matcher>[
-          isMethodCall('launch', arguments: <String, Object>{
-            'url': 'http://example.com/',
-            'useSafariVC': true,
-            'useWebView': false,
-            'enableJavaScript': false,
-            'enableDomStorage': false,
-            'universalLinksOnly': false,
-            'headers': <String, String>{},
-          })
-        ],
-      );
-    });
+//     test('launch', () async {
+//       await launcher.launch(
+//         'http://example.com/',
+//         useSafariVC: true,
+//         useWebView: false,
+//         enableJavaScript: false,
+//         enableDomStorage: false,
+//         universalLinksOnly: false,
+//         headers: const <String, String>{},
+//       );
+//       expect(
+//         log,
+//         <Matcher>[
+//           isMethodCall('launch', arguments: <String, Object>{
+//             'url': 'http://example.com/',
+//             'useSafariVC': true,
+//             'useWebView': false,
+//             'enableJavaScript': false,
+//             'enableDomStorage': false,
+//             'universalLinksOnly': false,
+//             'headers': <String, String>{},
+//           })
+//         ],
+//       );
+//     });
 
-    test('launch with headers', () async {
-      await launcher.launch(
-        'http://example.com/',
-        useSafariVC: true,
-        useWebView: false,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: false,
-        headers: const <String, String>{'key': 'value'},
-      );
-      expect(
-        log,
-        <Matcher>[
-          isMethodCall('launch', arguments: <String, Object>{
-            'url': 'http://example.com/',
-            'useSafariVC': true,
-            'useWebView': false,
-            'enableJavaScript': false,
-            'enableDomStorage': false,
-            'universalLinksOnly': false,
-            'headers': <String, String>{'key': 'value'},
-          })
-        ],
-      );
-    });
+//     test('launch with headers', () async {
+//       await launcher.launch(
+//         'http://example.com/',
+//         useSafariVC: true,
+//         useWebView: false,
+//         enableJavaScript: false,
+//         enableDomStorage: false,
+//         universalLinksOnly: false,
+//         headers: const <String, String>{'key': 'value'},
+//       );
+//       expect(
+//         log,
+//         <Matcher>[
+//           isMethodCall('launch', arguments: <String, Object>{
+//             'url': 'http://example.com/',
+//             'useSafariVC': true,
+//             'useWebView': false,
+//             'enableJavaScript': false,
+//             'enableDomStorage': false,
+//             'universalLinksOnly': false,
+//             'headers': <String, String>{'key': 'value'},
+//           })
+//         ],
+//       );
+//     });
 
-    test('launch force SafariVC', () async {
-      await launcher.launch(
-        'http://example.com/',
-        useSafariVC: true,
-        useWebView: false,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: false,
-        headers: const <String, String>{},
-      );
-      expect(
-        log,
-        <Matcher>[
-          isMethodCall('launch', arguments: <String, Object>{
-            'url': 'http://example.com/',
-            'useSafariVC': true,
-            'useWebView': false,
-            'enableJavaScript': false,
-            'enableDomStorage': false,
-            'universalLinksOnly': false,
-            'headers': <String, String>{},
-          })
-        ],
-      );
-    });
+//     test('launch force SafariVC', () async {
+//       await launcher.launch(
+//         'http://example.com/',
+//         useSafariVC: true,
+//         useWebView: false,
+//         enableJavaScript: false,
+//         enableDomStorage: false,
+//         universalLinksOnly: false,
+//         headers: const <String, String>{},
+//       );
+//       expect(
+//         log,
+//         <Matcher>[
+//           isMethodCall('launch', arguments: <String, Object>{
+//             'url': 'http://example.com/',
+//             'useSafariVC': true,
+//             'useWebView': false,
+//             'enableJavaScript': false,
+//             'enableDomStorage': false,
+//             'universalLinksOnly': false,
+//             'headers': <String, String>{},
+//           })
+//         ],
+//       );
+//     });
 
-    test('launch universal links only', () async {
-      await launcher.launch(
-        'http://example.com/',
-        useSafariVC: false,
-        useWebView: false,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: true,
-        headers: const <String, String>{},
-      );
-      expect(
-        log,
-        <Matcher>[
-          isMethodCall('launch', arguments: <String, Object>{
-            'url': 'http://example.com/',
-            'useSafariVC': false,
-            'useWebView': false,
-            'enableJavaScript': false,
-            'enableDomStorage': false,
-            'universalLinksOnly': true,
-            'headers': <String, String>{},
-          })
-        ],
-      );
-    });
+//     test('launch universal links only', () async {
+//       await launcher.launch(
+//         'http://example.com/',
+//         useSafariVC: false,
+//         useWebView: false,
+//         enableJavaScript: false,
+//         enableDomStorage: false,
+//         universalLinksOnly: true,
+//         headers: const <String, String>{},
+//       );
+//       expect(
+//         log,
+//         <Matcher>[
+//           isMethodCall('launch', arguments: <String, Object>{
+//             'url': 'http://example.com/',
+//             'useSafariVC': false,
+//             'useWebView': false,
+//             'enableJavaScript': false,
+//             'enableDomStorage': false,
+//             'universalLinksOnly': true,
+//             'headers': <String, String>{},
+//           })
+//         ],
+//       );
+//     });
 
-    test('launch force WebView', () async {
-      await launcher.launch(
-        'http://example.com/',
-        useSafariVC: true,
-        useWebView: true,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: false,
-        headers: const <String, String>{},
-      );
-      expect(
-        log,
-        <Matcher>[
-          isMethodCall('launch', arguments: <String, Object>{
-            'url': 'http://example.com/',
-            'useSafariVC': true,
-            'useWebView': true,
-            'enableJavaScript': false,
-            'enableDomStorage': false,
-            'universalLinksOnly': false,
-            'headers': <String, String>{},
-          })
-        ],
-      );
-    });
+//     test('launch force WebView', () async {
+//       await launcher.launch(
+//         'http://example.com/',
+//         useSafariVC: true,
+//         useWebView: true,
+//         enableJavaScript: false,
+//         enableDomStorage: false,
+//         universalLinksOnly: false,
+//         headers: const <String, String>{},
+//       );
+//       expect(
+//         log,
+//         <Matcher>[
+//           isMethodCall('launch', arguments: <String, Object>{
+//             'url': 'http://example.com/',
+//             'useSafariVC': true,
+//             'useWebView': true,
+//             'enableJavaScript': false,
+//             'enableDomStorage': false,
+//             'universalLinksOnly': false,
+//             'headers': <String, String>{},
+//           })
+//         ],
+//       );
+//     });
 
-    test('launch force WebView enable javascript', () async {
-      await launcher.launch(
-        'http://example.com/',
-        useSafariVC: true,
-        useWebView: true,
-        enableJavaScript: true,
-        enableDomStorage: false,
-        universalLinksOnly: false,
-        headers: const <String, String>{},
-      );
-      expect(
-        log,
-        <Matcher>[
-          isMethodCall('launch', arguments: <String, Object>{
-            'url': 'http://example.com/',
-            'useSafariVC': true,
-            'useWebView': true,
-            'enableJavaScript': true,
-            'enableDomStorage': false,
-            'universalLinksOnly': false,
-            'headers': <String, String>{},
-          })
-        ],
-      );
-    });
+//     test('launch force WebView enable javascript', () async {
+//       await launcher.launch(
+//         'http://example.com/',
+//         useSafariVC: true,
+//         useWebView: true,
+//         enableJavaScript: true,
+//         enableDomStorage: false,
+//         universalLinksOnly: false,
+//         headers: const <String, String>{},
+//       );
+//       expect(
+//         log,
+//         <Matcher>[
+//           isMethodCall('launch', arguments: <String, Object>{
+//             'url': 'http://example.com/',
+//             'useSafariVC': true,
+//             'useWebView': true,
+//             'enableJavaScript': true,
+//             'enableDomStorage': false,
+//             'universalLinksOnly': false,
+//             'headers': <String, String>{},
+//           })
+//         ],
+//       );
+//     });
 
-    test('launch force WebView enable DOM storage', () async {
-      await launcher.launch(
-        'http://example.com/',
-        useSafariVC: true,
-        useWebView: true,
-        enableJavaScript: false,
-        enableDomStorage: true,
-        universalLinksOnly: false,
-        headers: const <String, String>{},
-      );
-      expect(
-        log,
-        <Matcher>[
-          isMethodCall('launch', arguments: <String, Object>{
-            'url': 'http://example.com/',
-            'useSafariVC': true,
-            'useWebView': true,
-            'enableJavaScript': false,
-            'enableDomStorage': true,
-            'universalLinksOnly': false,
-            'headers': <String, String>{},
-          })
-        ],
-      );
-    });
+//     test('launch force WebView enable DOM storage', () async {
+//       await launcher.launch(
+//         'http://example.com/',
+//         useSafariVC: true,
+//         useWebView: true,
+//         enableJavaScript: false,
+//         enableDomStorage: true,
+//         universalLinksOnly: false,
+//         headers: const <String, String>{},
+//       );
+//       expect(
+//         log,
+//         <Matcher>[
+//           isMethodCall('launch', arguments: <String, Object>{
+//             'url': 'http://example.com/',
+//             'useSafariVC': true,
+//             'useWebView': true,
+//             'enableJavaScript': false,
+//             'enableDomStorage': true,
+//             'universalLinksOnly': false,
+//             'headers': <String, String>{},
+//           })
+//         ],
+//       );
+//     });
 
-    test('launch force SafariVC to false', () async {
-      await launcher.launch(
-        'http://example.com/',
-        useSafariVC: false,
-        useWebView: false,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: false,
-        headers: const <String, String>{},
-      );
-      expect(
-        log,
-        <Matcher>[
-          isMethodCall('launch', arguments: <String, Object>{
-            'url': 'http://example.com/',
-            'useSafariVC': false,
-            'useWebView': false,
-            'enableJavaScript': false,
-            'enableDomStorage': false,
-            'universalLinksOnly': false,
-            'headers': <String, String>{},
-          })
-        ],
-      );
-    });
+//     test('launch force SafariVC to false', () async {
+//       await launcher.launch(
+//         'http://example.com/',
+//         useSafariVC: false,
+//         useWebView: false,
+//         enableJavaScript: false,
+//         enableDomStorage: false,
+//         universalLinksOnly: false,
+//         headers: const <String, String>{},
+//       );
+//       expect(
+//         log,
+//         <Matcher>[
+//           isMethodCall('launch', arguments: <String, Object>{
+//             'url': 'http://example.com/',
+//             'useSafariVC': false,
+//             'useWebView': false,
+//             'enableJavaScript': false,
+//             'enableDomStorage': false,
+//             'universalLinksOnly': false,
+//             'headers': <String, String>{},
+//           })
+//         ],
+//       );
+//     });
 
-    test('closeWebView default behavior', () async {
-      await launcher.closeWebView();
-      expect(
-        log,
-        <Matcher>[isMethodCall('closeWebView', arguments: null)],
-      );
-    });
-  });
-}
+//     test('closeWebView default behavior', () async {
+//       await launcher.closeWebView();
+//       expect(
+//         log,
+//         <Matcher>[isMethodCall('closeWebView', arguments: null)],
+//       );
+//     });
+//   });
+// }
 
-class UrlLauncherPlatformMock extends Mock
-    with MockPlatformInterfaceMixin
-    implements UrlLauncherPlatform {}
+// class UrlLauncherPlatformMock extends Mock
+//     with MockPlatformInterfaceMixin
+//     implements UrlLauncherPlatform {}
 
-class ImplementsUrlLauncherPlatform extends Mock
-    implements UrlLauncherPlatform {}
+// class ImplementsUrlLauncherPlatform extends Mock
+//     implements UrlLauncherPlatform {}
 
-class ExtendsUrlLauncherPlatform extends UrlLauncherPlatform {}
+// class ExtendsUrlLauncherPlatform extends UrlLauncherPlatform {}
